Command manager update
It took me almost 2 entire days to make this.
BTW this is **NOT** finished yet.
The reason of why im uploading this is because i just finished the add command to list every time the bot runs (it was hard af) i cant even explain what really i have done. Just imagine this commit as a monkey finally finding the banana on a tree but not yet catching it.